+++ base/src/test/java/eu/fraho/spring/securityJwt/spring/UserDetailsServiceTestImpl.java
@@ -16,17 +16,14 @@ import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 
 import java.util.Collections;
-import java.util.Optional;
-import java.util.concurrent.atomic.AtomicInteger;
+import java.util.concurrent.atomic.AtomicBoolean;
 
 @Service
 public class UserDetailsServiceTestImpl implements UserDetailsService {
     public static final String BASE32_TOTP = "MZXW6YTBOI======";
-
+    public final transient AtomicBoolean apiAccessAllowed = new AtomicBoolean(true);
     @Autowired
     private PasswordEncoder passwordEncoder = null;
-
-    private AtomicInteger noRefreshCheckCount = new AtomicInteger(0);
 
     @Override
     public UserDetails loadUserByUsername(String username) throws UsernameNotFoundException {
@@ -42,14 +39,10 @@ public class UserDetailsServiceTestImpl implements UserDetailsService {
         } else {
             user.setAuthorities(Collections.singletonList(new SimpleGrantedAuthority("ROLE_USER")));
         }
-        if (username.equals("noRefresh")) {
-            user.setApiAccessAllowed(noRefreshCheckCount.getAndIncrement() % 2 == 1);
-        } else {
-            user.setApiAccessAllowed(true);
-        }
+        user.setApiAccessAllowed(apiAccessAllowed.get());
 
         if (username.equals("user_totp")) {
-            user.setTotpSecret(Optional.of(BASE32_TOTP));
+            user.setTotpSecret(BASE32_TOTP);
         }
 
         return user;