+++ hibernate/src/main/java/eu/fraho/spring/securityJwt/dto/RefreshTokenEntity.java
@@ -19,6 +19,7 @@ import java.time.ZonedDateTime;
 @Getter
 @Setter
 @EqualsAndHashCode(of = {"username", "deviceId", "token"})
+@ToString(of = {"id", "username", "deviceId"})
 @NoArgsConstructor
 public class RefreshTokenEntity {
     @Id
@@ -29,6 +30,7 @@ public class RefreshTokenEntity {
     @NotNull
     @Column(updatable = false)
     @Setter(AccessLevel.NONE)
+    @Convert(converter = ZonedDateTimeConverter.class)
     private ZonedDateTime created = ZonedDateTime.now();
 
     @NotNull