+++ hibernate/src/test/java/eu/fraho/spring/securityJwt/TestAuthControllerHibernate.java
@@ -8,19 +8,19 @@ package eu.fraho.spring.securityJwt;
 
 import eu.fraho.spring.securityJwt.controller.AbstractTestAuthControllerWithRefresh;
 import eu.fraho.spring.securityJwt.spring.TestHibernateApiApplication;
-import lombok.AccessLevel;
 import lombok.Getter;
-import lombok.Setter;
 import lombok.extern.slf4j.Slf4j;
 import org.junit.runner.RunWith;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
+import org.springframework.transaction.annotation.Transactional;
 
 @Getter
-@Setter(AccessLevel.NONE)
 @Slf4j
 @SpringBootTest(properties = "spring.config.location=classpath:hibernate-test.yaml",
         classes = TestHibernateApiApplication.class)
 @RunWith(SpringJUnit4ClassRunner.class)
+//@Commit
+@Transactional
 public class TestAuthControllerHibernate extends AbstractTestAuthControllerWithRefresh {
 }